move to marlin; clean-up
